+++ Assets/Prefabs/UI/InventoryPanel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 440432f5cf1ea6e44b5d9b60e8a9e4f4
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
