+++ Assets/Systems/Gameplay/Inventory/InventoryUI/ItemLabel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 59e3244129e9b1d43b1ccd72330cd5d6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
