+++ Assets/Systems/Gameplay/Inventory/InventoryUI/ItemLabel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7b77bd6651c554248a0a1fe195e17d84
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
