+++ Assets/Systems/Level Design/ResourceSpot.cs
@@ -16,5 +16,6 @@ public class ResourceSpot : MonoBehaviour
     public void GiveItemToPlayer(PlayerInventorySO inventory)
     {
         inventory.AddItemToInventory(resourceItem);
+        
     }
 }
