--- Assets/Systems/UI/InventoryUI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 87f676ffb5b074645abbcaed53a6c25c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
