--- Assets/Systems/UI/InventoryUI/InventoryUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9d8019e0909adad409552c10ee0ca990
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
