--- Assets/Systems/UI/InventoryUI/ItemIcon.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 59e3244129e9b1d43b1ccd72330cd5d6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
